requirements.txt
scikit-learn
tqdm
jupyterlab
tensorboard
pillow
seaborn
matplotlib
torch
torchvision
streamlit


